flightdeck update
renames esbuild2 and includes the legacy esbuild processing that uses the eleventy.after - keeping this file for the time being even thought we are not using it.
